Refactoring configuration into properties files Test cases not yet passing

+++ src/main/java/au/edu/wehi/idsv/metrics/SoftClipDetailMetrics.java
@@ -6,7 +6,7 @@ public class SoftClipDetailMetrics extends MetricBase {
 	/**
 	 * Length of soft clip
 	 */
-	public long LENGTH;
+	public int LENGTH;
 	/**
 	 * Number of reads with a soft clip of this length
 	 */
